Optimize: remove redundant judgment from SandHook
